+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Updated success messages for notification banners
+- Users who are not assigned to the project can no longer update the task list
+  actions
 
 ### Added